Improve imports and detailed consultation of users
The importation of users from CSV is expanded to support all fields of detail, including "antecedentes personales", "factores de riesgo", "antecedentes familiares" and "mujeres". A new endpoint '/users/detailed' is added to obtain all the consolidated information of users and their relationships.
Tests and examples of CSV files are updated to reflect the new supported fields.